misc/HtmlTestSuite/results/test-0998.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>paused state when removing from a document</title>
<link href="https://bugzilla.mozilla.org/show_bug.cgi?id=1583052" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/common/media.js"/>
</head><body><div id="log"/>
<div>
  <video hidden=""/>
</div>
<script>
function afterStableState(func) {
  var a = new Audio();
  a.volume = 0;
  a.addEventListener('volumechange', func);
}

async_test(function(t) {
  var v = document.querySelector('video');

  // Much like pause-remove-from-document.html, modulo this call.
  document.body.appendChild(v);

  v.src = getVideoURI('/media/movie_300');
  v.play();
  v.onplaying = t.step_func(function() {
    assert_false(v.paused, 'paused after playing');
    v.parentNode.removeChild(v);
    assert_false(v.paused, 'paused after removing');
    afterStableState(t.step_func(function() {
      assert_true(v.paused, 'paused after stable state');
      v.onpause = t.step_func(function() {
        assert_true(v.paused, 'paused in pause event');
        // re-insert and verify that it stays paused
        document.body.appendChild(v);
        t.step_timeout(function() {
          assert_true(v.paused, 'paused after re-inserting');
          t.done();
        }, 0);
      });
    }));
  });
});
</script>
</body></html>
